+++ crates/polars-plan/src/dsl/function_expr/boolean.rs
@@ -1,7 +1,3 @@
-use std::ops::Not;
-
-use polars_core::with_match_physical_integer_polars_type;
-
 use super::*;
 #[cfg(feature = "is_in")]
 use crate::wrap;
@@ -215,14 +211,5 @@ fn all_horizontal(s: &[Series]) -> PolarsResult<Series> {
 }
 
 fn not(s: &Series) -> PolarsResult<Series> {
-    match s.dtype() {
-        DataType::Boolean => Ok(s.bool().unwrap().not().into_series()),
-        dt if dt.is_integer() => {
-            with_match_physical_integer_polars_type!(dt, |$T| {
-                    let ca: &ChunkedArray<$T> = s.as_any().downcast_ref().unwrap();
-                    Ok(ca.apply_values(|v| !v).into_series())
-            })
-        },
-        dt => polars_bail!(InvalidOperation: "dtype {:?} not supported in 'not' operation", dt),
-    }
+    polars_ops::series::not(s)
 }
